# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KG-MNP Eligibility API",
    "description": "Frontend-ready backend for deterministic MNP eligibility assessment.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/assessments": {
      "post": {
        "tags": [
          "assessments"
        ],
        "summary": "Create Assessment",
        "operationId": "create_assessment_api_v1_assessments_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      },
      "get": {
        "tags": [
          "assessments"
        ],
        "summary": "List Assessments",
        "operationId": "list_assessments_api_v1_assessments_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Case Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentListResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/assessments/compare": {
      "get": {
        "tags": [
          "assessments"
        ],
        "summary": "Compare Assessments",
        "operationId": "compare_assessments_api_v1_assessments_compare_get",
        "parameters": [
          {
            "name": "left",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Left"
            }
          },
          {
            "name": "right",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Right"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/assessments/{execution_id}": {
      "get": {
        "tags": [
          "assessments"
        ],
        "summary": "Get Assessment",
        "operationId": "get_assessment_api_v1_assessments__execution_id__get",
        "parameters": [
          {
            "name": "execution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Execution Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentRecordResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/assessments/{execution_id}/trace": {
      "get": {
        "tags": [
          "assessments"
        ],
        "summary": "Get Trace",
        "operationId": "get_trace_api_v1_assessments__execution_id__trace_get",
        "parameters": [
          {
            "name": "execution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Execution Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/assessments/{execution_id}/artifacts": {
      "get": {
        "tags": [
          "assessments"
        ],
        "summary": "Get Artifacts",
        "operationId": "get_artifacts_api_v1_assessments__execution_id__artifacts_get",
        "parameters": [
          {
            "name": "execution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Execution Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactsResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/assessments/{execution_id}/what-if": {
      "post": {
        "tags": [
          "assessments"
        ],
        "summary": "Assessment What If",
        "operationId": "assessment_what_if_api_v1_assessments__execution_id__what_if_post",
        "parameters": [
          {
            "name": "execution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Execution Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatIfRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhatIfResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtifactsResponse": {
        "properties": {
          "execution_id": {
            "type": "string",
            "title": "Execution Id"
          },
          "artifacts": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Artifacts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "execution_id"
        ],
        "title": "ArtifactsResponse"
      },
      "AssessmentCreateRequest": {
        "properties": {
          "payload": {
            "$ref": "#/components/schemas/MNPCaseInput"
          },
          "persist": {
            "type": "boolean",
            "title": "Persist",
            "default": true
          },
          "force_recompute": {
            "type": "boolean",
            "title": "Force Recompute",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "payload"
        ],
        "title": "AssessmentCreateRequest"
      },
      "AssessmentListResponse": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AssessmentListResponse"
      },
      "AssessmentRecordResponse": {
        "properties": {
          "execution_id": {
            "type": "string",
            "title": "Execution Id"
          },
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "assessment_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Time"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssessmentResponse"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "input": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "execution_id",
          "case_id"
        ],
        "title": "AssessmentRecordResponse"
      },
      "AssessmentResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "execution_id": {
            "type": "string",
            "title": "Execution Id"
          },
          "case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "assessment_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Time"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EligibilityDecisionCode"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "publication": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublicationResponse"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Publication"
          },
          "validations": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidationsResponse"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Validations"
          },
          "input_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Summary"
          },
          "evidence": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EvidenceResponse"
                },
                {
                  "additionalProperties": true,
                  "type": "object"
                }
              ]
            },
            "type": "array",
            "title": "Evidence"
          },
          "rule_results": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RuleResultResponse"
                },
                {
                  "additionalProperties": true,
                  "type": "object"
                }
              ]
            },
            "type": "array",
            "title": "Rule Results"
          },
          "blocking_reasons": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BlockingReasonResponse"
                },
                {
                  "additionalProperties": true,
                  "type": "object"
                }
              ]
            },
            "type": "array",
            "title": "Blocking Reasons"
          },
          "remediation_actions": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RemediationActionResponse"
                },
                {
                  "additionalProperties": true,
                  "type": "object"
                }
              ]
            },
            "type": "array",
            "title": "Remediation Actions"
          },
          "process": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProcessResponse"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Process"
          },
          "trace_subgraph": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TraceGraphResponse"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Trace Subgraph"
          },
          "inference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InferenceResponse"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Inference"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "artifacts": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Artifacts"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "schema_version",
          "execution_id",
          "publication",
          "validations"
        ],
        "title": "AssessmentResponse"
      },
...
      "ComparisonResponse": {
        "properties": {
          "decision_changed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Changed"
          },
          "before": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before"
          },
          "after": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "After"
          },
          "added_blocking_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Added Blocking Reasons"
          },
          "removed_blocking_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Removed Blocking Reasons"
          },
          "changed_rule_versions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Changed Rule Versions"
          },
          "changed_evidence": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              }
            ],
            "title": "Changed Evidence"
          },
          "rule_changes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Rule Changes"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ComparisonResponse"
      },
...
      "EligibilityDecisionCode": {
        "type": "string",
        "enum": [
          "ELIGIBLE",
          "BLOCKED",
          "CONDITIONAL",
          "MANUAL_REVIEW"
        ],
        "title": "EligibilityDecisionCode"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "WhatIfRequest": {
        "properties": {
          "changes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Changes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "WhatIfRequest"
      },
...
    }
  }
}